1. init to Normal mode 2. PowerSaving <---> Normal <---> Setup: mode switch rule (by pinP1 limitation) 3. readme update

+++ examples/lora_ex1/lora_ex1.ino
@@ -203,7 +203,7 @@ boolean parseCommand(char *strCmd)
         {
           if (0 == i) {
             Serial.print("frequency: ");
-            freq = atoi(strOut);
+            freq = atol(strOut);
             Serial.print(freq);
           }
           else if (1 == i) {
